Implement the FILTER_VALIDATE_URL functionality.
This is funny. I didn't want to have to parse urls myself so I looked at
the "parse_url" function. In the docs is specifically says not to use that
function for validation. Well that sucks. So I dug into the source of the
extension to figure out how they did it. Guess what they use? Yeah, that's
right: parse_url. So basically this branch checks for a non zero length string,
runs parse url on it. Assuming that parse_url runs correctly the function
will return the input appropriately. If the additional flags are set the output
of parse_url is checked for the appropriate data.